refactor: move controller from folfer
